requirements.txt
django
djangorestframework
markdown
django-filter
mysqlclient
pyyaml
requests
django-cors-headers
python-decouple